+++ nautilus_trader/model/instruments/options_spread.pyx
@@ -79,8 +79,6 @@ cdef class OptionsSpread(Instrument):
     Raises
     ------
     ValueError
-        If `underlying` is not a valid string.
-    ValueError
         If `strategy_type` is not a valid string.
     ValueError
         If `multiplier` is not positive (> 0).
@@ -111,7 +109,6 @@ cdef class OptionsSpread(Instrument):
         str exchange = None,
         dict info = None,
     ):
-        Condition.valid_string(underlying, "underlying")
         Condition.valid_string(strategy_type, "strategy_type")
         Condition.positive_int(multiplier, "multiplier")
         if exchange is not None: